requirements.txt
anthropic>=0.49.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
streamlit>=1.43.0
